restructure directories for Poetry
