+++ README.md
@@ -61,7 +61,7 @@ This command will create the virtual environment and install all the necessary p
 Once installed, you can interact with the Jockey directly:
 
 ```bash
-poetry run python jockey.py --help
+poetry run jockey --help
 ```
 
 ### Hacking with Jockey
@@ -77,38 +77,38 @@ You will need to do this each time you start a new session.
 #### Performing checks locally
 ```bash
 # Lint import order with isort
-poetry run isort . --check
+poetry run isort src --check
 
 # Fix import order with isort
-poetry run isort .
+poetry run isort src
 
 # Lint with flake8
-poetry run flake8 .
+poetry run flake8 src
 
 # Lint with mypy
-poetry run mypy .
+poetry run mypy src
 
 # Lint code format with black
-poetry run black . --check
+poetry run black src --check
 
 # Fix code format with black
-poetry run black .
+poetry run black src
 
 # Execute unit tests
 poetry run pytest -s
 ```
 
 ## Examples:
-- `./jockey.py units`
-- `./jockey.py units charm=ceph-osd`
-- `./jockey.py charms machine=1`
-- `./jockey.py app charm=charm-nrpe machine=4/lxd/2`
-- `./jockey.py units app^=nova principal=true hostname~blrt`
+- `jockey units`
+- `jockey units charm=ceph-osd`
+- `jockey charms machine=1`
+- `jockey app charm=charm-nrpe machine=4/lxd/2`
+- `jockey units app^=nova principal=true hostname~blrt`
 
 ## Wishlist:
 
 * Allow comma-delimited selection of multiple object types:
-    `$ ./jockey.py get unit,app`
+    `jockey get unit,app`
 * Filter by leadership
 * Filter by principal
 * Filter by lxd
