+++ pyproject.toml
@@ -13,6 +13,13 @@ classifiers = [
     "Operating System :: POSIX :: Linux",
 ]
 
+packages = [
+    { include = "jockey", from = "src" }
+]
+
+[tool.poetry.scripts]
+jockey = "jockey.__main__:main"
+
 [tool.poetry.dependencies]
 python = ">=3.9,<3.13"
 click = "^8.1.7"
@@ -36,6 +43,12 @@ show_missing = true
 
 [tool.coverage.html]
 directory = "htmlcov"
+
+[tool.pytest.ini_options]
+pythonpath = "src"
+addopts = [
+    "--import-mode=importlib",
+]
 
 [tool.black]
 line-length = 120
@@ -100,7 +113,7 @@ version = "6.7.0"
 disable-upx = true
 
 [tool.poetry-pyinstaller-plugin.scripts]
-jockey = { source = "./jockey.py", type = "onefile", bundle = false }
+jockey = { source = "src/jockey/__main__.py", type = "onefile", bundle = false }
 
 [build-system]
 requires = ["poetry-core"]
