+++ src/jockey/__init__.py
@@ -1,0 +1,11 @@
+"""Jockey is a utility that facilitates quick retrieval of Juju objects that match given filters."""
+
+__author__ = "Connor Chamberlain"
+__credits__ = ["Connor Chamberlain", "Pedro Castillo", "John Lettman"]
+__maintainer__ = "Connor Chamberlain"
+__email__ = "connor.chamberlain@canonical.com"
+__version__ = "0.1.0"
+
+# TODO: Add the license when #28 is closed
+__license__ = "TBD"
+__copyright__ = "TBD"
