+++ tests/test_object_types.py
@@ -1,4 +1,4 @@
-from jockey import convert_object_abbreviation, ObjectType
+from src.jockey import ObjectType, convert_object_abbreviation
 
 
 def test_object_type_retrieval():
